It loads but won't allocate THAT much memory
However... It's not calling the IOMMU DMA ops so *maybe* I need to
allocate an IOMMU domain first and attach the driver to it.

Can this be done when we're making a random character device driver
though...

I guess I should continue from https://github.com/Miouyouyou/Nyanbada
code base.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>