Is 0x00000000 a valid IOVA ?
Well, technically, yes. If the IOMMU has no virtual mappings and can
start at 0, then the first mapping will return 0.

This is *EXTREMELY* disturbing when you're used to CPU Virtual
addresses and tends to consider that :
if (address == 0) { The_API_is_buggy(); }

...

I might just allocate a random element rounded to the IOMMU PAGE_SIZE
so that all addresses are different from 0... for debugging purposes.

Anyway, the registers are now almost set.
There's still one register where the Rockhip developpers went like :
Hey ! Let's just pack the frame size in the registers with the
File Descriptor !
So, now, I need to replicate their behaviour.
And, yes, RKMPP doesn't provide the addresses of the allocated
buffers, but the "File Descriptors" that correspond to DMA buffers
previously allocated on the GPU.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>