Fixed the MMAP issue...
And discovered that the VPU does a run but outputs nothing...
The output is completely untouched...

The IRQ is triggered and reg[1] shows : 0x00010100
According to the only documentation I have, this means that the
IRQ is active and the IRQ is disabled at the same time.
However the error bit is not set...

Ugh...

The clocks are setup and running at appropriate speeds
(aclk : 200 Mhz, iface : 50 Mhz).

Maybe I forgot some clocks... But only the HEVC DTS node is declared
with 4 clocks. The simple H264 decoder should work with only two
active...
Or maybe the VPU doesn't like 0x00000000 as a DMA address, even though
0x00000000 is a valid IOVA, as stated previously.

Hmm...

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>